+++ controllers/frontControllers.js
@@ -16,6 +16,11 @@ const getHome = async (req, res) => {
     }
 };
 
+// *api/logout - salir de la app
+//api/signin - registrarse en la app
+// hacer func
+
+
 
 
 const getSearch = async (req, res) => {
@@ -114,6 +119,72 @@ const getScrap = async (req, res) => {
     }
 }
 
+//"changePassword": `UPDATE users SET password = $1 WHERE users.email = $2`
+const restorePassword = async (req, res, next) => {
+    let client;
+    const { password, email } = req.query
+    try {
+        client = await pool.connect();
+        const data = await client.query(userQueries.changePassword, [password, email])
+        return res.status(200).json("Password cambiado correctamente")
+    } catch (err) {
+        return next(err);
+    } finally {
+        client.release();
+    }
+}
+
+
+//"recoverPassword":`SELECT password FROM users WHERE users.email = $1`
+const recoverPassword = async (req, res, next) => {
+    let client, result;
+    const { email } = req.query
+    try {
+        client = await pool.connect();
+        const data = await client.query(userQueries.recoverPassword, [email])
+        result = data.rows
+        return res.status(200).json(result)
+    } catch (err) {
+        return next(err);
+    } finally {
+        client.release();
+    }
+}
+
+
+const saveFavorite = async (req, res, next) => {
+    let client;
+    const { body } = req
+
+    try {
+        client = await pool.connect();
+        const data = await pool.query(userQueries.saveFav)
+        //no existe query - falta pasarle los datos
+
+        return res.status(200).json("Favorito guardado correctamente")
+    } catch (err) {
+        return next(err);
+    } finally {
+        client.release();
+    }
+}
+
+//`DELETE FROM favorites WHERE reference_offer = $1`
+const deleteFavorite = async (req, res, next) => {
+    let client;
+    const { reference_offer } = req.params
+    try {
+        client = await pool.connect();
+        const data = await pool.query(userQueries.deleteFav, [reference_offer])
+        return res.status(200).json(`Favorito con reference_offer ${reference_offer} eliminado correctamente`)
+    } catch (err) {
+        return next(err);
+    } finally {
+        client.release();
+    }
+
+}
+
 module.exports = {
     getHome,
     getSingup,
@@ -124,5 +195,8 @@ module.exports = {
     getDashboard,
     getScrap,
     getSearch,
-
+    restorePassword,
+    recoverPassword,
+    saveFavorite,
+    deleteFavorite
 };
