+++ routes/frontRoutes.js
@@ -1,6 +1,5 @@
 const express = require('express')
 const frontController = require("../controllers/frontControllers");
-const offerController = require("../controllers/offerControllers");
 const frontRouter = express.Router();
 
 frontRouter.get('/', frontController.getHome);
@@ -8,13 +7,18 @@ frontRouter.get('/singup', frontController.getSingup);
 frontRouter.get('/login', frontController.getLogin);
 frontRouter.get('/favorites', frontController.getFavorites);
 frontRouter.get('/profile', frontController.getProfile);
+
 //Prueba scraping
 frontRouter.post('/search', frontController.getSearch)
 frontRouter.get('/search', frontController.getSearch)
+
 //Vistas solo de admin
 frontRouter.get('/users', frontController.getUsers);
 frontRouter.get('/dashboard', frontController.getDashboard);
 
+//estas dos no llevan "/api" delante
+router.get('/recoverpassword', frontController.recoverPassword);
+router.get('/restorepassword', frontController.restorePassword);
 
 
 module.exports = frontRouter;
